+++ NEWS.rst
@@ -6,6 +6,7 @@ of every change, see the Git log.
 
 Latest
 ------
+* Minor: Added exceptions for accidental empty options. 
 * Major: Full rewrite of our Waf depedency resolve code.
 * Minor: Support for new resolver options.
 * Minor: Adding support for `resolve.json` files.